refactor: divide package for logger
