+++ gateway/logger/log_option.go
@@ -1,4 +1,4 @@
-package gateway
+package logger
 
 import (
 	"net/http"
