+++ gateway/logger/log_pattern_parser.go
@@ -1,4 +1,4 @@
-package gateway
+package logger
 
 import (
 	"os"
