+++ gateway/logger/logger_test.go
@@ -1,8 +1,8 @@
-package gateway_test
+package logger_test
 
 import (
 	"encoding/csv"
-	"gateway"
+	"gateway/logger"
 	"net/http"
 	"os"
 	"testing"
@@ -30,15 +30,15 @@ func TestUpdateLog(t *testing.T) {
 	r, _ := http.NewRequest(http.MethodGet, "http://example.com", nil)
 	r.Header.Set("TEST1", "header1")
 	r.Header.Set("TEST2", "header2")
-	gateway.LogOptionPattern = []gateway.LogOption{
-		gateway.WithTime(),
-		gateway.WithKey(),
-		gateway.WithPath(),
-		gateway.HeaderElement("TEST1"),
-		gateway.HeaderElement("TEST2"),
+	logger.LogOptionPattern = []logger.LogOption{
+		logger.WithTime(),
+		logger.WithKey(),
+		logger.WithPath(),
+		logger.HeaderElement("TEST1"),
+		logger.HeaderElement("TEST2"),
 	}
 	for i := 0; i < 2; i++ {
-		gateway.UpdateLog("key", "path", r)
+		logger.UpdateLog("key", "path", r)
 	}
 
 	// check if log is valid
